# internal/adapter/http/docs/swagger.yaml
basePath: /api
definitions:
...
  http.categoryResponse:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  http.expenseResponse:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      due_date:
        type: string
      expense_date:
        type: string
      group_id:
        type: string
      id:
        type: string
      installment_count:
        type: integer
      is_installment:
        type: boolean
      split_type:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
    type: object
  http.expenseSplitWithUserResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      expense_id:
        type: string
      id:
        type: string
      is_paid:
        type: boolean
      user_email:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
...
  http.installmentResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      due_date:
        type: string
      expense_id:
        type: string
      id:
        type: string
      installment_number:
        type: integer
      is_paid:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: luis@example.com
    name: Luis Mackiewicz
  description: Backend for the Doheem expense management application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Luis-Mackiewicz/doheem-backend
  title: Doheem API
  version: "1.0"
paths:
...
  /api/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an expense category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: group_id query param required
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Expenses
    put:
      consumes:
      - application/json
      description: Update an existing expense category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category update details
        in: body
        name: request
        required: true
        schema:
          properties:
            group_id:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.categoryResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Expenses
  /api/expenses/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete an expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Expense not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an expense
      tags:
      - Expenses
    get:
      consumes:
      - application/json
      description: Get a single expense by its unique identifier
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.expenseResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Expense not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get expense by ID
      tags:
      - Expenses
    put:
      consumes:
      - application/json
      description: Update an existing expense's details
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.expenseResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Expense not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an expense
      tags:
      - Expenses
  /api/expenses/{id}/installments:
    get:
      consumes:
      - application/json
      description: List all installments for a specific installment-based expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.installmentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Expense not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List expense installments
      tags:
      - Expenses
  /api/expenses/{id}/splits:
    get:
      consumes:
      - application/json
      description: List all payment splits for a specific expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.expenseSplitWithUserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Expense not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List expense splits
      tags:
      - Expenses
  /api/expenses/installments/{id}/pay:
    patch:
      consumes:
      - application/json
      description: Mark an expense installment as paid
      parameters:
      - description: Installment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Installment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark installment as paid
      tags:
      - Expenses
  /api/expenses/splits/{id}/pay:
    patch:
      consumes:
      - application/json
      description: Mark an expense split as paid
      parameters:
      - description: Split ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Split not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark split as paid
      tags:
      - Expenses
...
  /api/groups/{groupId}/categories:
    get:
      consumes:
      - application/json
      description: List all expense categories for a specific group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.categoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Expenses
    post:
      consumes:
      - application/json
      description: Create a new expense category for a group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      - description: Category name
        in: body
        name: request
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.categoryResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - Expenses
  /api/groups/{groupId}/expenses:
    get:
      consumes:
      - application/json
      description: List all expenses for a specific group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.expenseResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List expenses by group
      tags:
      - Expenses
    post:
      consumes:
      - application/json
      description: Create a new expense with splits in a group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.expenseResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an expense
      tags:
      - Expenses
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"